refactor: LottoModel의 메소드들을 AppController로 이동
Model과 Controller의 역할 범위 수정
