+++ src/js/models/Lotto/Lotto.js
@@ -1,9 +1,19 @@
-import { getLottoNumberList } from '../../utils/lottoUtils.js';
-import { cloneObject } from '../../utils/utils.js';
+import { getRandomInt, cloneObject } from '../../utils/utils.js';
+import { LOTTO } from '../../configs/contants.js';
 
 export default class Lotto {
+  static getLottoNumber() {
+    return getRandomInt(LOTTO.NUMBER_RANGE.MIN, LOTTO.NUMBER_RANGE.MAX);
+  }
+
+  static getLottoNumberList() {
+    return Array(LOTTO.NUMBER_LENGTH)
+      .fill()
+      .map(() => Lotto.getLottoNumber());
+  }
+
   constructor() {
-    this.numbers = getLottoNumberList();
+    this.numbers = Lotto.getLottoNumberList();
 
     Object.freeze(this);
   }
